+++ tests/testthat.R
@@ -1,5 +1,6 @@
+Sys.setenv("OMP_THREAD_LIMIT" = 2)
+
 library(testthat)
 library(r2dii.match)
 
-Sys.setenv("OMP_THREAD_LIMIT" = 2)
 test_check("r2dii.match")