+++ NEWS.md
@@ -1,6 +1,6 @@
 # r2dii.match (development version)
 
-* @jacobvjk is now the maintainer
+* @jacobvjk is now the maintainer.
 
 # r2dii.match 0.3.0