+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * `data_dictionary` dataset added to define the columns in each dataset used or exported by the functions in this package
 
+* @jacobvjk is now the maintainer.
+
 # r2dii.match 0.3.0
 
 ## Breaking changes