+++ NEWS.md
@@ -1,4 +1,4 @@
-# r2dii.match (development version)
+# r2dii.match 0.4.0
 
 * `data_dictionary` dataset added to define the columns in each dataset used or exported by the functions in this package